pyproject.toml
[build-system]
build-backend = "maturin"
requires = ["maturin>=1.0,<2.0"]

[project]
name="spiht-py"
version="0.1.1"
description = "SPIHT image compression algorithm with a rust backend."
dependencies = [
  "colour-science == 0.4.4",
  "numpy",
  "einops",
  "imageio",
  "matplotlib",
  "Pillow",
  "tqdm",
  "PyWavelets==1.5.0",
]
